backend/requirements.txt
Django
django-ninja
djangorestframework-simplejwt
